requirements.txt
Flask                    # the web framework itself
flask-cors               # to enable CORS on your API
Flask-Caching            # in-memory caching for sheets/API hits (per MVP1)  
gspread                  # Google Sheets connector  
google-auth              # for Sheets API auth  
pandas                   # data manipulation  
requests                 # Harvest API calls  
python-dotenv            # load .env secrets  
pytest                   # testing framework
Flask-Caching
requests
openai
redis